video: use int base type to store video fps rational number
Was double (numerator AND denominator types!): too high for nothing.
And wrong as coming from AVRational, made with int.

Tuleap: #237

+++ src/media/media_decoder.h
@@ -87,7 +87,7 @@ class MediaDecoder {
 
         int getWidth() const;
         int getHeight() const;
-        rational<double> getFps() const;
+        rational<int> getFps() const;
         int getPixelFormat() const;
 
         void setOptions(const std::map<std::string, std::string>& options);
